refactor: make acbs authentication api tests re-usable
